+++ CHANGELOG.md
@@ -4,14 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased - 2021-11-18
+## Unreleased - 2022-01-12
+### Added
+ * Double quotes on lookups can be escaped by adding `"escape_quotes": true` to the config file.
+
 ### Fixed
 * Fix and skip some non-working tests.
 * `Lookup.send_headers` method not working as expected with key_index parameter.
 * Add `type_of_key` parameter to `Lookup.send_headers` method.
 * Avoid lookups' key element to be deleted when `Lookup.send_data_line` is invoked.
 * **api-query**: return `None` when the query response has no results. 
-
 
 ## [3.4.2] - 2021-05-18
 ### Added